axionch-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
cryptography==43.0.1
httpx==0.27.2
tweepy==4.14.0
Pillow==10.4.0
argon2-cffi==23.1.0
